Remove excess progress bar and edited output
I removed one of the progress bar options because I think they were stacking when only one was supposed to be used. I edited the output as well so that it would be on a new line and not off the page after the progress bar. I also added a confirmation from the user to close the script because if you run the file by itself it will close before you can see the output.